requirements.txt
requests==2.21.0
python-dotenv==0.10.1
vk_api==11.4.0
python-telegram-bot==11.1.0
